Added primary and mini styles to both tab components
